move some to azure/core
