+++ packages/cadl-ranch-specs/http/azure/core/main.cadl
@@ -33,6 +33,25 @@ model User {
 
   @doc("The user's name.")
   name: string;
+
+  @doc("The user's order list")
+  orders?: UserOrder[];
+  ...global.Azure.Core.EtagProperty;
+}
+
+@doc("UserOrder for testing list with expand.")
+@resource("user")
+model UserOrder {
+  @key
+  @doc("The user's id.")
+  @visibility("read")
+  id: int32;
+
+  @doc("The user's id.")
+  userId: int32;
+
+  @doc("The user's order detail")
+  detail: string;
 }
 
 @doc("The parameters for exporting a user.")
@@ -139,7 +158,22 @@ Expected response body:
 }
 ```
 """)
-op list is ResourceOperations.ResourceList<User>;
+op list is ResourceOperations.ResourceList<
+  User,
+  ListQueryParametersTrait<global.Azure.Core.StandardListQueryParameters &
+    global.Azure.Core.OrderByQueryParameter &
+    global.Azure.Core.FilterQueryParameter &
+    global.Azure.Core.SelectQueryParameter &
+    global.Azure.Core.ExpandQueryParameter>
+>;
+
+#suppress "@azure-tools/cadl-azure-core/use-standard-operations" "For testing global.Azure.Core.Page"
+@scenario
+@doc("List with Azure.Core.Page<>.")
+@route("/page")
+@scenarioDoc("""
+  """)
+op listWithPage is global.Azure.Core.Foundations.Operation<{}, global.Azure.Core.Page<User>>;
 
 @scenario
 @doc("Deletes a User")
