+++ packages/cadl-ranch-specs/http/azure/custom-by-traits/mockapi.ts
@@ -1,13 +1,7 @@
-import { passOnSuccess, mockapi, json, ValidationError, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, mockapi, json, ValidationError } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
-
-function validateHeaders(req: MockRequest) {
-  if (!("x-ms-client-request-id" in req.headers)) {
-    throw new ValidationError("Should submit header x-ms-client-request-id", "any uuid", undefined);
-  }
-}
 
 const validUser = {
   id: 1,
@@ -15,26 +9,11 @@ const validUser = {
   ETag: "11bdc430-65e8-45ad-81d9-8ffa60d55b59",
 };
 
-Scenarios.Azure_Traits_create = passOnSuccess(
-  mockapi.post("/azure/traits/user", (req) => {
-    validateHeaders(req);
-    const validBody = { name: "Madge" };
-    req.expect.bodyEquals(validBody);
-    return {
-      status: 201,
-      body: json(validUser),
-      headers: {
-        "Location": "http://localhost:3000/azure/traits/user/1",
-        "Repeatability-Result": "Accepted",
-        "x-ms-client-request-id": req.headers["x-ms-client-request-id"],
-      },
-    };
-  }),
-);
-
 Scenarios.Azure_Traits_get = passOnSuccess(
   mockapi.get("/azure/traits/user/:id", (req) => {
-    validateHeaders(req);
+    if (!("x-ms-client-request-id" in req.headers)) {
+      throw new ValidationError("Should submit header x-ms-client-request-id", "any uuid", undefined);
+    }
     if (req.params.id !== "1") {
       throw new ValidationError("Expected path param id=1", "1", req.params.id);
     }
@@ -55,42 +34,11 @@ Scenarios.Azure_Traits_get = passOnSuccess(
   }),
 );
 
-Scenarios.Azure_Traits_list = passOnSuccess(
-  mockapi.get("/azure/traits/user", (req) => {
-    validateHeaders(req);
-    req.expect.containsQueryParam("top", "5");
-    req.expect.containsQueryParam("skip", "10");
-    req.expect.containsQueryParam("orderby", "id");
-    req.expect.containsQueryParam("filter", "id eq 1");
-    if (!req.originalRequest.originalUrl.includes("select=id&select=orders&select=ETag")) {
-      throw new ValidationError("Expected query param colors=blue&colors=red&colors=green ", "1", req.headers["select"]);
-    }
-    req.expect.containsQueryParam("expand", "orders");
-    const responseBody = {
-      value: [
-        {
-          id: 1,
-          ETag: "11bdc430-65e8-45ad-81d9-8ffa60d55b59",
-          orders: [{ id: 1, userId: 1, detail: "a recorder" }],
-        },
-      ],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-);
-
-Scenarios.Azure_Traits_listWithPage = passOnSuccess(
-  mockapi.get("/azure/traits/page", (req) => {
-    const responseBody = {
-      value: [validUser],
-    };
-    return { status: 200, body: json(responseBody) };
-  }),
-);
-
 Scenarios.Azure_Traits_delete = passOnSuccess(
   mockapi.delete("/azure/traits/api/2022-12-01-preview/user/:id", (req) => {
-    validateHeaders(req);
+    if (!("x-ms-client-request-id" in req.headers)) {
+      throw new ValidationError("Should submit header x-ms-client-request-id", "any uuid", undefined);
+    }
     if (req.params.id !== "1") {
       throw new ValidationError("Expected path param id=1", "1", req.params.id);
     }
